+++ src/components/2020/Contact/css/index.module.scss
@@ -1,0 +1,23 @@
+.contact{
+  &_title{
+    margin-bottom: 1.5em;
+  }
+
+  &_control_margin{
+    margin-left: 1em;
+  }
+
+  &_column {
+    padding: 3rem;
+
+    @media all and (max-width: 767.99px){
+      padding: 0;
+    }
+    &:nth-child(1){
+      order: 2;
+    }
+    &:nth-child(2){
+      order: 1;
+    }
+  }
+}
