+++ src/components/2020/GoodFirms/css/index.module.scss
@@ -1,0 +1,8 @@
+.good_firms_container{
+  position: relative;
+  padding-top: 2em;
+}
+
+.image{
+  width: 85%;
+}
