+++ src/components/2020/Testimonial.js
@@ -10,6 +10,7 @@ const Testimonial = () => {
     <Media className="component-changes">
       <MediaLeft
         className="testimonial-image"
+        
         style={{
           backgroundImage: `url(${image})`,
         }}
