+++ src/components/enterprise/containers/Menu.js
@@ -51,6 +51,7 @@ const ToggleContainer = styled.span`
   }
 `;
 
+
 const MenuTitle = styled.span`
   color: #264A60;
   line-height: 40px;
